Market Potential Scraper/requirements.txt
# Web 自动化
playwright==1.40.0

# 数据处理
pandas==2.1.4
openpyxl==3.1.2

# 环境变量管理
python-dotenv==1.0.0

# HTTP 请求（可选，用于 API 调用）
requests==2.31.0

# AI 翻译（可选）
openai==1.6.1
google-generativeai==0.3.2
